Big Integer and addition of help target in make script
I already had the code for the Make script, and fixing bigint on the tables was an easy fix. You can decide on whether to keep making script or not.  On windows with ODBC, they won’t work well.  Have a blessed day, Stefan!

+++ database/tables/checksums.sql
@@ -1,6 +1,6 @@
 CREATE TABLE checksums (
-    id SERIAL PRIMARY KEY,
-    file_id INTEGER NOT NULL REFERENCES files(id),
+    id BIGSERIAL PRIMARY KEY,
+    file_id bigint NOT NULL REFERENCES files(id),
     checksum_sha256 TEXT NOT NULL,
     checksum_sha3 TEXT NOT NULL,
     checksum_blake_2b TEXT NOT NULL
